chore(logo): move files
