--- packages/design-system/src/components/Logo/index.ts
@@ -1,2 +1,0 @@
-export { default as Lunit } from "./Lunit";
-export { default as LunitTrademark } from "./LunitTrademark";
